Refactor

+++ src/main/java/com/example/blegarden/BindingAdapters.kt
@@ -8,7 +8,8 @@ import com.example.blegarden.enums.PositionsInArray
 
 @BindingAdapter(value = ["app:onOrOff", "app:intOfArray", "app:deviceName"])
 fun onOrOff(button: Button, byteArray: ByteArray, positionInArray: Int, deviceName: String) {
-    button.text = if (byteArray[positionInArray] == 0.toByte()) "$deviceName OFF" else "$deviceName ON"
+    button.text =
+        if (byteArray[positionInArray] == 0.toByte()) "$deviceName OFF" else "$deviceName ON"
 }
 
 @BindingAdapter("app:startOrStop")
@@ -23,6 +24,6 @@ fun visibleOrZero(view: View, dataArray: ByteArray) {
 
 
 @BindingAdapter(value = ["app:setMode", "app:modeStatus"])
-fun setMode(radioButton: RadioButton, dataArray: ByteArray, mode: Int) {
-    radioButton.isChecked = dataArray[PositionsInArray.MODE_STATUS.position] == mode.toByte()
+fun setMode(radioButton: RadioButton, dataArray: ByteArray, mode: Byte) {
+    radioButton.isChecked = dataArray[PositionsInArray.MODE_STATUS.position] == mode
 }
